Slatter/requirements.txt
dj-database-url==0.5.0
Django==4.0.2
django-cors-headers==3.11.0
djangorestframework==3.13.1
entrypoints==0.3
filelock==3.6.0
greenlet==1.1.2
gunicorn==20.1.0
httplib2==0.14.0
idna==2.8
importlib-metadata==4.11.3
lockfile==0.12.2
oauthlib==3.1.0
olefile==0.46
paramiko==2.6.0
psycopg2==2.9.3
psycopg2-binary==2.9.3
reportlab==3.5.34
requests==2.22.0
SQLAlchemy==1.4.35
sqlparse==0.4.2
ssh-import-id==5.10
systemd-python==234
virtualenv==20.14.0
